core-components: document LogViewer + restructure
